Fixed a problem in a test
The test wich caused cbc to abort, compared the obtained lower_bound
against a hardcoded value, which was known to be the optimal solution.
However, this test failed sometimes at Travis. The only explanation is
that the cbc version used at Travis was different, and due to rounding
errors the optimal solution was also different.

Now the test computes both the optimal (solving without time limit) and
the lower bound (solving with tight time limit) using the same solver,
so the assertion lower_bound <= optimal must hold true. Otherwise there
is an error to debug somewhere.