malloovia package is a facade to the modules
In `__init__.py` all important classes and methos are imported from
the submodules, and thus exported by `malloovia` namespace. This allows
the user to do `from malloovia import *` and have access to these
classe, or alternatively `import malloovia` and have access to
`malloovia.PhaseI`, `malloovia.read_problem_from_yaml()`, etc. as if all
those methods and classes were implemented in a single package.

To avoid making too much information available this way, each submodule
declares in the attribute `__all__` the list of identifiers to export
when `from .module import *` is used.

This does not restrict the access to one particular method in one
particular submodule if the user knows the implementation details, as
for example `from malloovia.util import _dict_to_yaml`. It does not
simplify the autogenerated documenation, either (except that private
methods, which start by _, are not included by default in the
documentation generated by Sphinx).