Fixed: default solver is not COIN
phases.py forced the solver to be COIN() if the user did not specify
one. However, this would cause PhaseI.solve() and PhaseII.solve_period()
to fail in machines in which COIN-OR CBC is not installed.

Now no solver is forced, so the default's one which comes with PuLP is
used, and thus malloovia can work in platforms in which COIN-OR CBC is
not installed. Note however that one of the tests uses COIN solver to
specify a maxSeconds, so that test will fail in those machines.