Fixed: broken build in travis
Travis installs mallovia by running `pip install .`, which runs
`setup.py`. In the latest commit, `setup.py` imported `malloovia`
to get the version number from `malloovia.__version__`, but this
import failed because it required `pulp` which was not installed
at that time.

To break this dependency, a new file `malloovia/__version__.py` is
included. Malloovia imports this file during normal operation, and
exports the `__version__` attribute it contains, so the API was not
changed. But `setup.py` does not import malloovia anymore, but instead
reads the `__version__` directly from that file. This technique was
copied from `requests` library.