Fix problems in the schema
The "OneOf" of Solutions should be "oneOf" (it did not correctly validate
before) and some required properties in phase_ii_solution were missing.

In the generation of optimal_cost for solving_stats, "None", "True" and
"False" were changed to valid json: "null", "true" and "false".

Additionally, a test for an aborted solution with "null" in the
optimal_cost has been added.