Added time_unit in the model
This required several changes:
* The schema now *requires* the property time_unit in Instance_classes,
  Workloads and PerformanceSet elements. This is a string which can take
  the value "y", "h", "m" or "s"
* model.py also requires this property for the corresponding classes
* util.py was modified to include that property when objects of these
  classes are instantiated, and to generate these fields in the
  to_yaml functions.
* The .yaml examples were modified to add this property where required
* The tests were rewritten to add this property when programatically
  building problems.

In all the examples (yaml files and programatically created ones), the
property time_unit is set to the value "h".

The solver currently ignores this property, so it still behaves as if
all those elements were using the same unit. But, at least, all the
tests still pass.