Fixed several type errors
Now the whole malloovia repository pass `mypy` without errors

You have to specify `--no-strict-optional` to `mypy`, because some
fields could be `None`, under some circunstances. For example, if the
LP problem has no solution, the fields `optimal_cost`, `allocation` and
others of the `SolutionI` class could be `None`.

This introduces other possible bugs when trying to operate in those
fields without asserting first that they are not `None`. Dealing
properly with this issue would require deep changes in the codebase.

For now, it is simpler to use `--no-strict-optional` (which was indeed
the default option in previous versions of `mypy`)