Bugfix: Return valid objects in solutions_to_yaml()
Fixes #16

Now `solutions_from_dict()` to which `read_solutions_from_yaml()`
delegates, returns proper SolutionI and SolutionII objects.

**However**, `SolutionII.allocation.workload_tuples` can be an empty
list if the yaml from which the solution is read does not include this
field (because in the yaml schema this field is optional). This makes
sense to avoid the duplication of the Workload in the solution, since it
is already in the problem. However, in order to provide auto-sufficient
allocation information, perhaps the code should populate that list.

The test was also changed to use a new yaml example file, namely
`tests/test_data/valid/examples_documentation-sol.yaml`. which is
an _actual_ solution written by Malloovia (as oposed to the previous
example file which as hand-written as an example and did not contain
a valid solution). The test is augmented to verify some basic properties
of the solution (type of the fields, length of the allocation, etc.)