Bugfix: anchors are now always valid in yaml
The anchors used in the YAML files generated by functions
`problems_to_yaml` and `solutions_to_yaml` use now automatically
generated anchors, instead of using the `.id` field of the objects,
since this field could contain invalid characters for an anchor (eg:
spaces, commas, unicode, etc. ) Moreover, the `.id` fields could be
duplicated, causing problem in anchors.

Now each anchor is generated by concatenating a "sanitized" version of
the `.id` field (if present), obtained by replacing each invalid char by
`_`, and an hexadecimal constant which is guaranteed to be unique for
each object (implementation detail: is the `id()` of the object).

This fixes #15.

To implement this feature several changes were neccessary:

* A function `_anchor_from_id()` was added to `util.py` * Most of the
functions in `util.py` which generated yaml were modified to use this
function. * Some objects in `util.py` were modified "on the fly", which
caused their `id()` to change (because they were inmutable namedtuple
objects, and then the modifications via `._replace()` created new
tuples). To avoid this problem the anchor was created before the
modifications.

A new test was added which abuses of invalid chars in ids.