Bugfix: Solutions are written and read correctly
The code of `malloovia.util` has been rewritten to ensure that the
generated yaml uses lists instead of tuples (because yaml/json has
not the tuple type), but it is converted back to tuples when read.

Also, several data structures which were returned as dicts, are
now returning the proper types (SolvingStats, MalloviaStats,
GlobalSolvingStats)

The test `test_arbitrary_id_produce_valid_anchors()` has been
extended to test also the proper working of `util.solutions_to_yaml()`
and `util.solutions_from_dict()` which are the ones used to write
(resp. read) the YAML files.

This fixes #17