Shorter LP variable names
The names of the LP variables are automatically created from the repr
of several Python objects, one of these being a tuple with the
workload values for all the apps in a timeslot. When the number of
apps is large, the ascii repr of that tuple is large too, which can
cause problems to the LP sover because the variable names exceed
some limit.

To solve this issue, the histogram uses as keys a custom tuple object
whose repr has fixed length (it is the hash of the tuple).

A new test was added to check that using a large number of apps
doesn't cause problems.

Fixes #19